refactor(primitives): finish ts_ms storage — pitchfork + legacy migration + dead-helper cleanup
- Complete ts_ms coord refactor for remaining primitives: signals, projections (long/short position, bars_pattern, price_projection, projection), volumes (fixed/anchored), pitchfork family (inside_pitchfork), fibonacci fan
- Add #[serde(default)] to all 77 tsN/anchor_ts fields across 37 primitive files for legacy preset compatibility
- Implement DrawingManager::migrate_legacy_timestamps_for_primitive and migrate_all_legacy_timestamps — converts point_timestamps[i]*1000 → ts_ms when ts fields are all zero (old Unix-seconds format)
- Wire migration into both preset load paths in chart_out_events.rs
- Fix Command::CreatePrimitive/DeletePrimitive/MovePrimitive/RestoreAll/DeleteAll points type: Vec<(f64,f64)> → Vec<(i64,f64)>
- Fix UndoAction::PrimitiveMoved old/new_points: Vec<(f64,f64)> → Vec<(i64,f64)>
- Fix drag_start_points field: Option<(usize, Vec<(f64,f64)>)> → Vec<(i64,f64)>
- Delete dead helpers: ensure_timestamps_populated, recalculate_all_bar_caches, update_all_timestamps_from_bars calls across tick.rs, modals.rs, lib.rs, mod.rs, settings.rs, chart_window.rs
- Fix all caller type mismatches: hit_test(+bars), on_click(ts_ms), start_drag(ts_ms), start_control_point_drag(ts_ms), update_drag(ts_ms), translate_at(ts_delta_ms), set_points(&[(i64,f64)])
- Fix alert boundary: convert ts_ms→bar_f64 when passing drawing_points to alerts crate (which still uses bar-index internally)
- Add bars param to draw_alert_bell_icons for ts_ms→bar_f64 conversion in alert rendering
- Fix signal_manager tests: float literal timestamps → i64 Unix ms literals
- Fix agent commands.rs: point arrays mapped as (p[0] as i64, p[1])
- cargo check --all-targets: 0 errors, 0 warnings

+++ crates/chart-app-vello/src/agent/commands.rs
@@ -167,7 +167,7 @@ impl App<'_> {
                     window_id, chart_id, type_id, points, style, agent_id: _,
                 } => {
                     if let Some(pw) = self.windows.values_mut().find(|pw| pw.window_id == window_id) {
-                        let pts: Vec<(f64, f64)> = points.iter().map(|p| (p[0], p[1])).collect();
+                        let pts: Vec<(i64, f64)> = points.iter().map(|p| (p[0] as i64, p[1])).collect();
                         let color_str = style.color.clone();
                         let registry = zengeld_chart::drawing::primitives_v2::PrimitiveRegistry::global().read().unwrap();
                         if let Some(mut prim) = registry.create(&type_id, &pts, Some(&color_str)) {
@@ -205,7 +205,7 @@ impl App<'_> {
                                 .iter_mut().find(|p| p.data().id == primitive_id);
                             if let Some(prim) = prim {
                                 if let Some(pts) = points {
-                                    let new_pts: Vec<(f64, f64)> = pts.iter().map(|p| (p[0], p[1])).collect();
+                                    let new_pts: Vec<(i64, f64)> = pts.iter().map(|p| (p[0] as i64, p[1])).collect();
                                     prim.set_points(&new_pts);
                                 }
                                 if let Some(s) = style {
